allow hours worked to be decimal

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisuserrolesapi/data/StaffJobClassification.kt
@@ -1,5 +1,6 @@
 package uk.gov.justice.digital.hmpps.nomisuserrolesapi.data
 
+import java.math.BigDecimal
 import java.time.LocalDate
 
 data class StaffJobClassification(
@@ -8,7 +9,7 @@ data class StaffJobClassification(
   val role: String,
   val position: String,
   val scheduleType: String,
-  val hoursPerWeek: Int,
+  val hoursPerWeek: BigDecimal,
   val fromDate: LocalDate,
   val toDate: LocalDate?,
 )
